requirements.txt
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.5
idna==3.6
pycparser==2.21
pycryptodome==3.20.0
pyOpenSSL==24.1.0
requests==2.31.0
urllib3==2.2.1
